split JetStream logic out of subscriber.rs into its own jetstream module
