+++ lib/llm/src/discovery/model_manager.rs
@@ -419,9 +419,8 @@ impl ModelManager {
         let instance_id = discovery.instance_id();
 
         // Build transport for router endpoint based on request plane mode
-        // Derive the router's discovery component name from the target worker component
-        let router_endpoint_id =
-            router_endpoint_id(endpoint.id().namespace, &endpoint.id().component);
+        // Use KV_ROUTER_COMPONENT as the component name to distinguish from the generate endpoint's component
+        let router_endpoint_id = router_endpoint_id(endpoint.id().namespace);
         let transport = build_transport_type(endpoint, &router_endpoint_id, instance_id).await?;
 
         let discovery_spec = DiscoverySpec::Endpoint {
