+++ lib/llm/src/kv_router.rs
@@ -29,6 +29,7 @@ pub use dynamo_kv_router::indexer;
 pub use dynamo_kv_router::protocols;
 
 pub mod config;
+mod jetstream;
 pub mod metrics;
 pub mod prefill_router;
 pub mod publisher;
@@ -88,26 +89,23 @@ pub fn worker_kv_indexer_query_endpoint(dp_rank: DpRank) -> String {
 }
 
 // for router discovery registration
+pub const KV_ROUTER_COMPONENT: &str = "kv-router";
 pub const KV_ROUTER_ENDPOINT: &str = "generate";
 
-fn kv_router_component_name(target_component: &str) -> String {
-    format!("{target_component}-kv-router")
-}
-
 /// Creates an EndpointId for the KV router in the given namespace.
-pub fn router_endpoint_id(namespace: String, target_component: &str) -> EndpointId {
+pub fn router_endpoint_id(namespace: String) -> EndpointId {
     EndpointId {
         namespace,
-        component: kv_router_component_name(target_component),
+        component: KV_ROUTER_COMPONENT.to_string(),
         name: KV_ROUTER_ENDPOINT.to_string(),
     }
 }
 
 /// Creates a DiscoveryQuery for the KV router in the given namespace.
-pub fn router_discovery_query(namespace: String, target_component: &str) -> DiscoveryQuery {
+pub fn router_discovery_query(namespace: String) -> DiscoveryQuery {
     DiscoveryQuery::Endpoint {
         namespace,
-        component: kv_router_component_name(target_component),
+        component: KV_ROUTER_COMPONENT.to_string(),
         endpoint: KV_ROUTER_ENDPOINT.to_string(),
     }
 }
